refactor(trip): flatten providers/ dir into providers.dart
Every other feature exposes its Riverpod surface via a single
providers.dart at the feature root. trip was the only outlier with a
providers/ directory that held just one file, so promote it to match
the convention and rewrite the 15 import sites.
