+++ frontend/lib/features/trip/presentation/screens/trip_edit_screen.dart
@@ -1,5 +1,5 @@
 import 'package:context_app/features/trip/domain/models/trip.dart';
-import 'package:context_app/features/trip/providers/trip_providers.dart';
+import 'package:context_app/features/trip/providers.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
