+++ frontend/lib/features/trip/presentation/screens/trip_list_screen.dart
@@ -1,6 +1,6 @@
 import 'package:context_app/app/config/app_colors.dart';
 import 'package:context_app/features/trip/presentation/widgets/trip_grid.dart';
-import 'package:context_app/features/trip/providers/trip_providers.dart';
+import 'package:context_app/features/trip/providers.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
