+++ frontend/lib/features/trip/providers.dart
