+++ frontend/test/features/trip/presentation/screens/trip_list_screen_test.dart
@@ -4,7 +4,7 @@ import 'package:context_app/features/trip/domain/models/trip.dart';
 import 'package:context_app/features/trip/presentation/screens/trip_edit_screen.dart';
 import 'package:context_app/features/trip/presentation/screens/trip_list_screen.dart';
 import 'package:context_app/features/trip/presentation/widgets/trip_card.dart';
-import 'package:context_app/features/trip/providers/trip_providers.dart';
+import 'package:context_app/features/trip/providers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
