+++ frontend/test/features/trip/presentation/widgets/move_to_trip_sheet_test.dart
@@ -1,5 +1,5 @@
 import 'package:context_app/features/trip/presentation/widgets/move_to_trip_sheet.dart';
-import 'package:context_app/features/trip/providers/trip_providers.dart';
+import 'package:context_app/features/trip/providers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
